+++ Kitodo/src/main/java/org/kitodo/production/forms/DesktopForm.java
@@ -253,22 +253,6 @@ public class DesktopForm extends BaseForm {
     }
 
     /**
-     * Retrieve comments of given process.
-     *
-     * @param processDTO
-     *          process for which comment tooltip is created and returned
-     * @return List containing comments for given process
-     */
-    public List<Comment> getComments(ProcessDTO processDTO) {
-        try {
-            return ServiceManager.getProcessService().getComments(processDTO);
-        } catch (DAOException e) {
-            Helper.setErrorMessage(e);
-            return Collections.emptyList();
-        }
-    }
-
-    /**
      * Check and return whether the process with the ID 'pid' has any correction comments or not.
      *
      * @param pid
